+++ openbas-api/src/main/java/io/openbas/utils/AgentUtils.java
@@ -1,6 +1,6 @@
 package io.openbas.utils;
 
-import io.openbas.database.model.Endpoint;
+import io.openbas.database.model.*;
 import java.util.List;
 
 public class AgentUtils {
@@ -17,4 +17,36 @@ public class AgentUtils {
       List.of(
           Endpoint.PLATFORM_ARCH.x86_64.name().toLowerCase(),
           Endpoint.PLATFORM_ARCH.arm64.name().toLowerCase());
+
+  public static List<Agent> getActiveAgents(Asset asset, Inject inject) {
+    return ((Endpoint) asset)
+        .getAgents().stream().filter(agent -> isValidAgent(inject, agent)).toList();
+  }
+
+  public static boolean isValidAgent(Inject inject, Agent agent) {
+    return isPrimaryAgent(agent) && hasOnlyValidTraces(inject, agent) && agent.isActive();
+  }
+
+  public static boolean hasOnlyValidTraces(Inject inject, Agent agent) {
+    return inject
+        .getStatus()
+        .map(InjectStatus::getTraces)
+        .map(
+            traces ->
+                Boolean.valueOf(
+                    traces.stream()
+                        .noneMatch(
+                            trace ->
+                                trace.getAgent() != null
+                                    && trace.getAgent().getId().equals(agent.getId())
+                                    && (ExecutionTraceStatus.ERROR.equals(trace.getStatus())
+                                        || ExecutionTraceStatus.AGENT_INACTIVE.equals(
+                                            trace.getStatus())))))
+        .orElse(Boolean.TRUE)
+        .booleanValue(); // If there are no traces, return true by default
+  }
+
+  public static boolean isPrimaryAgent(Agent agent) {
+    return agent.getParent() == null && agent.getInject() == null;
+  }
 }
